# Optimize Base.decode*